+++ packages/asset-service/src/generateAssetData/generateAssetData.ts
@@ -6,7 +6,7 @@ import orderBy from 'lodash/orderBy'
 
 import { Asset, AssetsById } from '../service/AssetService'
 import * as avalanche from './avalanche'
-import { atom, bitcoin } from './baseAssets'
+import { atom, bitcoin, dogecoin } from './baseAssets'
 import * as ethereum from './ethereum'
 import * as osmosis from './osmosis'
 import { setColors } from './setColors'
@@ -20,6 +20,7 @@ const generateAssetData = async () => {
   // all assets, included assets to be blacklisted
   const unfilteredAssetData: Asset[] = [
     bitcoin,
+    dogecoin,
     atom,
     ...ethAssets,
     ...osmosisAssets,